requirements.txt
# ===================================
# QRGenie Backend Dependencies
# ===================================

# Django Core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
django-extensions==3.2.3
dj-database-url==2.3.0

# Authentication
djangorestframework-simplejwt==5.4.0
PyJWT==2.9.0
pyotp==2.9.0

# Celery (async tasks)
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.1

# File Storage
boto3==1.35.0
django-storages==1.14.4
minio==7.2.12

# QR Code Generation
qrcode[pil]==8.0
Pillow==11.0.0
reportlab==4.1.0
segno==1.6.5

# Excel / Bulk Operations
openpyxl==3.1.5
xlsxwriter==3.2.0

# AI / OpenAI
openai==1.58.1
groq==0.13.0

# GeoIP
geoip2==4.8.1
maxminddb==2.6.2

# HTTP Client
httpx==0.28.1
requests==2.32.3

# Utilities
python-decouple==3.8
nanoid==2.0.0
bleach==6.2.0
python-slugify==8.0.4
hashids==1.3.1
bcrypt==4.2.1

# User-Agent Parsing
ua-parser==1.0.0

# Server
gunicorn==23.0.0

# Development
django-debug-toolbar==4.4.6
ipython==8.30.0
